--- ide/gwt-logger/src/main/java/org/eclipse/che/ide/logger/logging/gwt/Logging.gwt.xml
@@ -1,37 +1,0 @@
-<!--
-
-    Copyright (c) 2012-2017 Codenvy, S.A.
-    All rights reserved. This program and the accompanying materials
-    are made available under the terms of the Eclipse Public License v1.0
-    which accompanies this distribution, and is available at
-    http://www.eclipse.org/legal/epl-v10.html
-
-    Contributors:
-      Codenvy, S.A. - initial API and implementation
-
--->
-<module>
-  <inherits name="org.eclipse.che.ide.logger.logging.impl.LoggingImpl"/>
-  <inherits name="com.google.gwt.logging.Logging"/>
-
-  <source path=""/>
-
-
-  <replace-with class="org.eclipse.che.ide.logger.logging.gwt.GWTLoggerFactory">
-    <when-type-is class="org.slf4j.ILoggerFactory"/>
-    <none>
-      <when-property-is name="gwt.logging.enabled" value="FALSE"/>
-    </none>
-  </replace-with>
-
-
-  <set-configuration-property name="devModeRedirectEnabled" value="true"/>
-  <set-property name="compiler.useSourceMaps" value="true" />
-
-  <set-property name="user.agent" value="safari"/>
-  <set-configuration-property name="user.agent.runtimeWarning" value="false"/>
-
-  <set-property name="compiler.stackMode" value="emulated"/>
-  <set-configuration-property name="compiler.emulatedStack.recordLineNumbers" value="true"/>
-  <set-configuration-property name="compiler.emulatedStack.recordFileNames" value="true"/>
-</module>
