--- ide/gwt-logger/src/main/java/org/eclipse/che/ide/logger/logging/impl/LoggingImpl.gwt.xml
@@ -1,17 +1,0 @@
-<!--
-
-    Copyright (c) 2012-2017 Codenvy, S.A.
-    All rights reserved. This program and the accompanying materials
-    are made available under the terms of the Eclipse Public License v1.0
-    which accompanies this distribution, and is available at
-    http://www.eclipse.org/legal/epl-v10.html
-
-    Contributors:
-      Codenvy, S.A. - initial API and implementation
-
--->
-<module>
-    <inherits name="org.eclipse.che.ide.logger.api.API"/>
-
-    <source path=""/>
-</module>
