+++ ide/gwt-logger/src/main/java/org/eclipse/che/ide/logger/slf4j/emulation/org/slf4j/helpers/FormattingTuple.java
