+++ ide/gwt-logger/src/main/resources/org/eclipse/che/ide/logger/slf4j/Logging.gwt.xml
@@ -1,0 +1,32 @@
+<!--
+
+    Copyright (c) 2012-2017 Codenvy, S.A.
+    All rights reserved. This program and the accompanying materials
+    are made available under the terms of the Eclipse Public License v1.0
+    which accompanies this distribution, and is available at
+    http://www.eclipse.org/legal/epl-v10.html
+
+    Contributors:
+      Codenvy, S.A. - initial API and implementation
+
+-->
+<module>
+    <inherits name="com.google.gwt.core.Core"/>
+    <inherits name="com.google.gwt.logging.Logging"/>
+
+    <super-source path="emulation"/>
+    <source path="gwtbackend"/>
+
+    <replace-with class="org.eclipse.che.ide.logger.slf4j.gwtbackend.GWTLoggerFactory">
+        <when-type-is class="org.slf4j.ILoggerFactory"/>
+        <none>
+            <when-property-is name="gwt.logging.enabled" value="FALSE"/>
+        </none>
+    </replace-with>
+
+    <set-configuration-property name="devModeRedirectEnabled" value="true"/>
+    <set-property name="compiler.useSourceMaps" value="true"/>
+    <set-property name="compiler.stackMode" value="emulated"/>
+    <set-configuration-property name="compiler.emulatedStack.recordLineNumbers" value="true"/>
+    <set-configuration-property name="compiler.emulatedStack.recordFileNames" value="true"/>
+</module>
